admin/pyproject.toml
[tool.poetry]
name = "web"
version = "0.1.0"
description = ""
authors = ["fots <facundotomatis@gmail.com>"]
readme = "README.md"
packages = [{include = "web", from = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.2.2"
python-dotenv = "^0.21.0"
Flask-SQLAlchemy = "^2.5.1"
psycopg2-binary = "^2.9.3"
WTForms = "^3.0.1"
passlib = "^1.7.4"
Flask-WTF = "^1.0.1"
Flask-Bootstrap = "^3.3.7.1"
fpdf2 = "^2.5.7"
Flask-Session = "^0.4.0"
email-validator = "^1.3.0"
PyJWT = "^2.6.0"
flask-cors = "^3.0.10"
qrcode = "^7.3.1"
flask-jwt-extended = "^4.4.4"
requests = "^2.28.2"
pytest = "^7.2.1"
pytest-html = "^3.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
black = "^22.10.0"


[[tool.poetry.source]]
name = "fpdf"
url = "https://github.com/reingart/pyfpdf"
default = false
secondary = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
